Feat(#47): saveReview 수정

+++ src/main/java/com/lec/spring/matzip/repository/TagRepository.java
@@ -9,7 +9,7 @@ public interface TagRepository {
 
     Tag findById(int id);
 
-    List<Tag> findByIds(List<Integer> ids);
+    List<Tag> findByIds(List<Long> ids);
 
     List<Tag> findAll();
